+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move Terraform code from `terraform/aws` to `terraform`
 - Move `HistoricalDiary::SourceMaterialGenerator` side-effect code to separate method
 - Use `day_page_link` partial on person page
+- Separate `DayPage` and `YearPage` into individual generators
 
 ### Fixed
 - De-duplicate tags listed on day page